Poprawki konwersji Ptype
Przeladowanie Ptype == T i Ptype != T, wszystkie konwersje
Ptype (rowniez te dotyczace const char[]) uzywaja teraz referencji.

+++ src/Datasaver.cpp
@@ -1,7 +1,7 @@
 #include "Datasaver.hpp"
 
 void send_data(Database* db, string task_name, int session_id, map<string, Ptype> session_data, map<string, Ptype> trial_data){
-  if((int)trial_data["trial"] == 0){
+  if(trial_data["trial"] == 0){
     string q("CREATE TABLE IF NOT EXISTS " + task_name + "_data "
              "(timestamp TIMESTAMP, session_id INT(11) NOT NULL, ");
     for(auto& x : trial_data)
